% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_predictions}
\alias{check_predictions}
\title{Check that a predictions table can be plotted}
\usage{
check_predictions(predictions)
}
\arguments{
\item{predictions}{the \code{predictions} element of a \code{fit_patch_model()} result}
}
\value{
\code{TRUE} invisibly; errors otherwise
}
\description{
Check that a predictions table can be plotted
}
\keyword{internal}
