% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{covariate_names}
\alias{covariate_names}
\title{Covariate variable names in a covariate object}
\usage{
covariate_names(env_dat)
}
\arguments{
\item{env_dat}{an \code{sf} POINT covariate object}
}
\value{
character vector of variable column names
}
\description{
Covariate variable names in a covariate object
}
\keyword{internal}
