% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{critical_t}
\alias{critical_t}
\title{The two-sided critical value}
\usage{
critical_t(level, df)
}
\arguments{
\item{level}{confidence level}

\item{df}{degrees of freedom}
}
\value{
the critical \code{t}, or \code{NA_real_}
}
\description{
The two-sided critical value
}
\keyword{internal}
