% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_copernicus_datasets}
\alias{default_copernicus_datasets}
\title{Default Copernicus datasets}
\usage{
default_copernicus_datasets()
}
\value{
a list of product/dataset/variable specs for \code{covariates.copernicus}
}
\description{
Monthly physical ocean variables from the global reanalysis: potential
temperature, salinity, and the two current components. Chlorophyll lives in a
separate biogeochemistry product and is left for the user to add, since which
BGC product is appropriate depends on the year range.
}
