% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prejoin.R
\name{drop_covariate}
\alias{drop_covariate}
\title{Remove a covariate from the product carrying it}
\usage{
drop_covariate(per_dataset, index, covariate)
}
\arguments{
\item{per_dataset}{a list of \code{sf} POINT objects}

\item{index}{which object to take it from}

\item{covariate}{the covariate to drop}
}
\value{
\code{per_dataset} without that column
}
\description{
Remove a covariate from the product carrying it
}
\keyword{internal}
