% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{join_covariate_sources}
\alias{join_covariate_sources}
\title{Join covariate sets from two products onto the same points}
\usage{
join_covariate_sources(x, y)
}
\arguments{
\item{x}{an \code{sf} POINT covariate object}

\item{y}{another \code{sf} POINT covariate object}
}
\value{
\code{x} with \code{y}'s variable columns joined on nearest point and date
}
\description{
Join covariate sets from two products onto the same points
}
\keyword{internal}
