% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{print.taupatch_ensemble}
\alias{print.taupatch_ensemble}
\title{Print an ensemble}
\usage{
\method{print}{taupatch_ensemble}(x, ...)
}
\arguments{
\item{x}{a \code{taupatch_ensemble}}

\item{...}{unused}
}
\value{
\code{x}, invisibly
}
\description{
Print an ensemble
}
