% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{projection_map}
\alias{projection_map}
\title{Build a leaflet map of a projection GeoTIFF}
\usage{
projection_map(geotiff, opacity = 0.8)
}
\arguments{
\item{geotiff}{path to a projection GeoTIFF written by \code{project_patch_model()}}

\item{opacity}{overlay opacity}
}
\value{
a \code{leaflet} map widget
}
\description{
Build a leaflet map of a projection GeoTIFF
}
\examples{
\dontrun{
result <- run_taupatch("inst/configs/mock_test.yaml")
projection_map(result$projections$geotiff[1])
}
}
