% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{write_or_return}
\alias{write_or_return}
\title{Write a plot to disk or return it}
\usage{
write_or_return(plot, path, width = 7, height = 5)
}
\arguments{
\item{plot}{a ggplot object}

\item{path}{where to write, or \code{NULL} to return the plot}

\item{width, height}{PNG dimensions in inches}
}
\value{
the plot, or \code{path} invisibly
}
\description{
Write a plot to disk or return it
}
\keyword{internal}
